add normalize features

+++ Core/Models/OhlcBar.cs
@@ -10,8 +10,8 @@
         public decimal Close { get; init; }
         public decimal TypicalPrice => (High + Low + Close) / 3;
         public int TickVolume { get; init; }
-        public decimal UpVolume { get; init; } // NEW - sum of up tick volumes
-        public decimal DownVolume { get; init; } // NEW - sum of down tick volumes
+        public int UpVolume { get; init; } // NEW - sum of up tick volumes
+        public int DownVolume { get; init; } // NEW - sum of down tick volumes
         public decimal AvgSpread { get; init; }
         public decimal MaxSpread { get; init; }
         public decimal MinSpread { get; init; }
